Split aset and avas
