+++ forte/CMakeLists.txt
@@ -239,6 +239,8 @@ mrdsrg-spin-integrated/mrdsrg_srg.cc
 mrdsrg-spin-integrated/three_dsrg_mrpt2.cc
 orbital-helpers/ao_helper.cc
 orbital-helpers/aosubspace.cc
+orbital-helpers/aset.cc
+orbital-helpers/avas.cc
 orbital-helpers/ci-no/ci-no.cc
 orbital-helpers/ci-no/mrci-no.cc
 orbital-helpers/fragment_projector.cc
@@ -246,7 +248,6 @@ orbital-helpers/iao_builder.cc
 orbital-helpers/localize.cc
 orbital-helpers/mp2_nos.cc
 orbital-helpers/mrpt2_nos.cc
-orbital-helpers/orbital_embedding.cc
 orbital-helpers/orbitaloptimizer.cc
 orbital-helpers/pao_builder.cc
 orbital-helpers/semi_canonicalize.cc
