+++ forte/forte.cc
@@ -26,17 +26,9 @@
  * @END LICENSE
  */
 
-#include <fstream>
-
-#include "ambit/tensor.h"
-
-#include "psi4/libpsi4util/process.h"
-
-#include "orbital-helpers/aosubspace.h"
-#include "orbital-helpers/orbital_embedding.h"
-#include "base_classes/forte_options.h"
-#include "base_classes/mo_space_info.h"
-#include "helpers/timer.h"
+// #include <fstream>
+// #include "ambit/tensor.h"
+// #include "psi4/libpsi4util/process.h"
 
 #include "version.h"
 #include "psi4/libpsi4util/PsiOutStream.h"
