+++ forte/pymodule.py
@@ -171,7 +171,7 @@ def energy_forte(name, **kwargs):
     elif data.options.get_str("INT_TYPE") == "PYSCF":
         data = ObjectsFromPySCF(kwargs.get("pyscf_obj"), options=kwargs).run(data)
     else:
-        data = ObjectsFromPsi4(**kwargs).run(data)  # <--- This is where AVAS runs
+        data = ObjectsFromPsi4(**kwargs).run(data)
 
     start = time.time()
 
@@ -203,11 +203,11 @@ def energy_forte(name, **kwargs):
         data = MCSCF(active_space_solver_type).run(data)
         energy = data.results.value("mcscf energy")
 
-
+    # Optionally, invoke embedding
     if data.options.get_bool("EMBEDDING"):
         data.mo_space_info = make_embedding_orbitals(data)
         data.ints = forte.make_ints_from_psi4(data.psi_wfn, data.options, data.scf_info, data.mo_space_info)
-    
+
     # Run a method
     if job_type == "NONE":
         psi4.core.set_scalar_variable("CURRENT ENERGY", energy)
